tumble_race_game/event_shipper/requirements.txt
pyobjc-framework-Quartz
inputs
boto3
pygame
pyaudio
websocket-client
